RavenDB-18675 Introducing new namespace for Fields.
