+++ src/Corax/IndexSearcher/IndexSearcher.MultiTermMatches.cs
@@ -1,13 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 using Corax.Queries;
 using Corax.Queries.MultiTermMatch;
 using Corax.Queries.MultiTermMatch.TermProviders;
 using Corax.Utils;
-using Sparrow;
 using Voron;
 using Sparrow.Server;
 using Voron.Data.CompactTrees;
