+++ src/Corax/IndexSearcher/IndexSearcher.cs
@@ -10,6 +10,7 @@ using Voron.Data.Containers;
 using Sparrow;
 using System.Runtime.Intrinsics.X86;
 using Corax.Analyzers;
+using Corax.Fields;
 using Corax.IndexEntry;
 using Corax.Pipeline;
 using Corax.Queries;
