+++ src/Corax/Queries/SuggestionTermProvider.cs
@@ -1,16 +1,13 @@
 ﻿using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Corax.Fields;
 using Corax.Pipeline;
-using Corax.Utils;
 using Sparrow;
 using Sparrow.Collections;
-using Sparrow.Server;
 using Sparrow.Server.Strings;
-
 using Voron;
 using static Corax.Constants;
