+++ src/Raven.Server/Documents/Queries/QueryBuilderHelper.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Globalization;
 using System.Runtime.CompilerServices;
 using Corax;
+using Corax.Fields;
 using Corax.Queries;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Indexes.Spatial;
