+++ test/FastTests/Corax/BoostingQuery.cs
@@ -4,8 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Corax;
+using Corax.Fields;
 using Corax.IndexEntry;
 using Corax.Queries;
+using Corax.Queries.MultiTermMatch;
 using Corax.Queries.MultiTermMatch.TermProviders;
 using FastTests.Voron;
 using Sparrow.Server;
