+++ test/FastTests/Corax/CoraxQueries.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using Corax.Queries;
 using Corax;
+using Corax.Fields;
 using Corax.IndexEntry;
 using FastTests.Voron;
 using Sparrow.Server;
