+++ test/FastTests/Corax/IndexEntryWriter.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Corax;
+using Corax.Fields;
 using Corax.IndexEntry;
 using FastTests.Voron;
 using Sparrow;
