+++ test/FastTests/Corax/IndexSearcher.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using Corax;
 using Corax.Analyzers;
+using Corax.Fields;
 using Corax.IndexEntry;
 using Corax.Pipeline;
 using Corax.Queries;
