+++ test/FastTests/Corax/OrderByMultiSorting.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Corax;
+using Corax.Fields;
 using Corax.IndexEntry;
 using Corax.Queries;
 using FastTests.Voron;
