+++ test/FastTests/Corax/RawCoraxFlag.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Corax;
 using Corax.Analyzers;
+using Corax.Fields;
 using Corax.IndexEntry;
 using FastTests.Voron;
 using Raven.Server.Documents.Indexes.Persistence.Corax;
